pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-report-generator"
version = "0.1.0"
description = "基于参考资料生成可追溯报告的本地工具与 AI skills 集合"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Workspace Maintainer" }
]
dependencies = [
  "pypdf>=5.3.0",
  "python-docx>=1.1.2",
  "python-pptx>=1.0.2",
  "openpyxl>=3.1.5",
  "beautifulsoup4>=4.12.3",
  "requests>=2.32.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5"
]

[project.scripts]
report-generator = "report_generator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
